Rework payment activation flow
